src/books/requirements.txt
backcall==0.2.0
decorator==4.4.2
Django==2.2.16
django-debug-toolbar==3.1
django-extensions==3.0.8
Faker==4.1.2
flake8==3.8.3
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==3.1.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==5.0.4
wcwidth==0.2.5
